Project/requirements.txt
dash==2.18.2
dash-bootstrap-components==1.6.0
pandas==2.2.3
plotly==5.24.1
scikit-learn==1.6.0
numpy==2.0.2
umap-learn==0.5.7